backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
python-dotenv==1.0.0
python-multipart==0.0.6

# Google Gemini (Recommended)
google-generativeai>=0.3.2

# RAG and Vector Search
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Optional: If you want to use other LLM providers
# httpx==0.25.1  # For HuggingFace/OpenAI REST APIs
# openai==1.3.0
# anthropic==0.7.0

# Testing Dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0  # For TestClient
